src/requirements.txt
cloudshell-orch-core>=2.0.0.0,<2.2.0.0
cloudshell-shell-core>=4,<5
cloudshell-automation-api>=9.1,<9.2
cloudshell-traffic==3.0.3

xenavalkyrie==1.4.1
